// apps/admin/src/api/generated/schemas/channelUser.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Fox Sphere API
 * REST API for the Fox Sphere Twitch bot platform
 * OpenAPI spec version: 1.0.0
 */

export interface ChannelUser {
  twitchId: string;
  username: string;
  coins: number;
  lvl: number;
  xp: number;
  lastXpAt: string;
  isMod: boolean;
  isFounder: boolean;
  isSubscriber: boolean;
  isPermanentVip: boolean;
  spinsCount: number;
  totalWin: number;
  totalLoss: number;
  xpThisWeek: number;
  hasTicket: boolean;
  isLuckyVip: boolean;
}